+++ packages/carousel/src/react/__specs__/storyshots.spec.js
@@ -4,6 +4,7 @@ import initStoryshots, {
 } from '@storybook/addon-storyshots'
 
 jest.mock('@pluralsight/ps-design-system-storybook-addon-center')
+jest.mock('@pluralsight/ps-design-system-storybook-addon-chaos')
 jest.mock('@pluralsight/ps-design-system-storybook-addon-theme')
 
 jest.mock('../../js/utils.js', () => ({