+++ packages/site/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [17.3.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@17.3.5...@pluralsight/ps-design-system-site@17.3.6) (2020-01-08)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
+## [17.3.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@17.3.4...@pluralsight/ps-design-system-site@17.3.5) (2020-01-08)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
 ## [17.3.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@17.3.3...@pluralsight/ps-design-system-site@17.3.4) (2019-12-31)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-site